feat: Refactor file structure in app module
